chore: Update service worker to notify clients of new update
This commit updates the service worker to include a new function, notifyClientsOfUpdate(), which sends a message to all clients (UI) when a new update is available. The function is called after updating the cache files in the updateCacheFiles() function. This ensures that clients are notified immediately when a new version of the app is available.